requirements.txt
contourpy==1.0.7
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.38.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.3
numpy==1.24.1
openpyxl==3.1.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyparsing==3.0.9
PyQt5==5.15.8
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
six==1.16.0
